Conflicts:
	envsetup.sh
	livecd/create_livecd.sh
	set_version.sh

+++ envsetup.sh
@@ -274,6 +274,35 @@ function mkiso()
     fi
 }
 
+function mkiso_debug()
+{
+    if [ $# -lt 3 ] ; then
+        echo You should execute this cmd with three param at least as follow:
+        echo "mkiso_debug xxx.iso OUTPATH APPPATH"
+        return 1
+    fi
+
+    if [ ! -d $2 ] ; then
+        echo You should make sure the OUTPATH $1 is a dir
+        return 1
+    fi
+
+    if [ ! -d $3 ] ; then
+        echo You should make sure the APPPATH $2 is a dir
+        return 1
+    fi
+
+    T=$(gettop)
+    if [ ! "$T" ]; then
+        echo "fail to locate the top of the tree.  Try setting TOP."
+        return 1
+    fi
+
+    sudo sh $T/build/debug/installkdump.sh $2 $3 || return 1
+    mkiso $1 || return 1
+    sudo sh $T/build/debug/uninstallkdump.sh $2 $3 || return 1
+}
+
 function m()
 {
     T=$(gettop)
@@ -533,7 +562,9 @@ function mrootbuilder()
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "dpkg-divert --rename --remove /sbin/initctl"
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "apt-get -y --force-yes clean"
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "rm -rf /tmp/*"
-    sudo chroot $OUT/out/squashfs-root /bin/bash -c "rm /etc/resolv.conf"
+    echo "nameserver 8.8.8.8">/tmp/resolv.conf
+    sudo cp /tmp/resolv.conf $OUT/out/squashfs-root/etc/resolv.conf
+    #sudo chroot $OUT/out/squashfs-root /bin/bash -c "rm /etc/resolv.conf"
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "umount -lf /proc"
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "umount -lf /sys"
     sudo chroot $OUT/out/squashfs-root /bin/bash -c "umount -lf /dev/pts"
@@ -554,7 +585,9 @@ function createlink()
     do
         if [ -f $i ] ; then
         dirpath=$(dirname $i)
-        sudo ln -sf "$i" $dirpath/libudev.so.0
+        if [ ! -e $dirpath/libudev.so.0 ] ; then
+            sudo ln -sf "$i" $dirpath/libudev.so.0
+        fi
         echo "create link succefull "
         break
         fi
@@ -574,6 +607,7 @@ function _mcos()
     IS4LENOVO=0
     IS4S3G=0
     IS4TEST=0
+    IS4DEBUG=0
     ISFROMSRC=0
     if [ -e $BUILDOSSTEP ] ; then
         BUITSTEP=`cat $BUILDOSSTEP`
@@ -593,6 +627,8 @@ function _mcos()
             IS4S3G=1
 	elif [ "$i" == "--test" ] ; then
 	    IS4TEST=1
+	elif [ "$i" == "--debug" ] ; then
+	    IS4DEBUG=1
 	elif [ "$i" == "--srcbuild" ] ; then
 	    ISFROMSRC=1
         else
@@ -646,7 +682,7 @@ function _mcos()
             echo 31 >$BUILDOSSTEP
 	    if [ $ISFROMSRC -eq 1 ] ; then
                 sudo cp $T/build/release/tmpfiles/mdm/mdm.conf $OUTPATH/squashfs-root/etc/mdm/
-                uninstalldeb "symbol-fonts account-plugin-facebook account-plugin-flickr account-plugin-google account-plugin-twitter alacarte appmenu-gtk appmenu-gtk3 appmenu-qt appmenu-qt5 apport apport-symptoms bamfdaemon banshee-extension-soundmenu bison cdparanoia cdrdao compiz compiz-core compiz-gnome compiz-plugins-default curl dconf-tools docbook-xsl flex freepats friends-facebook friends-twitter gir1.2-panelapplet-4.0 gir1.2-rb-3.0 gir1.2-unity-5.0 gnome-applets gnome-applets-data gnome-control-center gnome-control-center-data gnome-control-center-signon gnome-control-center-unity gnome-media gnome-session gnome-session-fallback gnome-user-guide gromit gstreamer0.10-gnomevfs hud humanity-icon-theme icoutils indicator-applet-complete indicator-appmenu indicator-datetime indicator-messages indicator-power indicator-printers indicator-session indicator-sound k3b k3b-data kate-data katepart kde-runtime kde-runtime-data kde-style-oxygen kde-window-manager kde-window-manager-common kdelibs-bin kdelibs5-data kdelibs5-plugins kdoctools kubuntu-debug-installer libattica0.4 libbamf3-1 libbison-dev libcompizconfig0 libdlrestrictions1 libencode-locale-perl libfile-listing-perl libfl-dev libflac++6 libfont-afm-perl libgnome-control-center1 libgnome-media-profiles-3.0-0 libgnome2-canvas-perl libgnome2-perl libgnome2-vfs-perl libgnomevfs2-extra libhtml-form-perl libhtml-format-perl libhtml-parser-perl libhtml-tagset-perl libhtml-tree-perl libhttp-cookies-perl libhttp-daemon-perl libhttp-date-perl libhttp-message-perl libhttp-negotiate-perl libibus-1.0-0 libio-socket-ssl-perl libk3b6 libkactivities-bin libkactivities-models1 libkactivities6 libkatepartinterfaces4 libkcddb4 libkcmutils4 libkde3support4 libkdeclarative5 libkdecorations4abi1 libkdecore5 libkdesu5 libkdeui5 libkdewebkit5 libkdnssd4 libkemoticons4 libkfile4 libkhtml5 libkidletime4 libkio5 libkjsapi4 libkjsembed4 libkmediaplayer4 libknewstuff3-4 libknotifyconfig4 libkntlm4 libkparts4 libkpty4 libkrosscore4 libktexteditor4 libkwineffects1abi4 libkwinglutils1abi1 libkwinnvidiahack4 libkworkspace4abi2 libkxmlrpcclient4 liblwp-mediatypes-perl liblwp-protocol-https-perl libmusicbrainz5-0 libmysqlclient18 libnepomuk4 libnepomukcore4abi1 libnepomukquery4a libnepomukutils4 libnet-http-perl libnet-ssleay-perl libntrack-qt4-1 libntrack0 libnux-4.0-0 libphonon4 libplasma3 libpolkit-qt-1-1 libpoppler-qt4-4 libqapt2 libqapt2-runtime libqca2 libqt4-qt3support libqt4-sql-mysql librhythmbox-core6 libsolid4 libsoprano4 libstreamanalyzer0 libstreams0 libthreadweaver4 libunity-core-6.0-5 libunity-misc4 libunity-webapps0 libvirtodbc0 libwww-perl libwww-robotrules-perl libxcb-damage0 libxml2-utils mysql-common nautilus nepomuk-core nepomuk-core-data notification-daemon ntrack-module-libnl-0 odbcinst odbcinst1debian2 oxygen-icon-theme phonon phonon-backend-gstreamer plasma-scriptengine-javascript python-zeitgeist python3-apport python3-dbus.mainloop.qt python3-distupgrade python3-problem-report python3-pyqt4 python3-sip python3-update-manager qapt-batch rhythmbox rhythmbox-data rhythmbox-mozilla rhythmbox-plugin-cdrecorder rhythmbox-plugin-zeitgeist rhythmbox-plugins rhythmbox-ubuntuone shared-desktop-ontologies soprano-daemon ubiquity-frontend-kde ubuntu-release-upgrader-core unity unity-asset-pool unity-common unity-lens-applications unity-lens-files unity-lens-friends unity-lens-music unity-lens-photos unity-lens-shopping unity-lens-video unity-scope-gdrive unity-scope-musicstores unity-scope-video-remote unity-services unity-webapps-service update-manager-core vcdimager virtuoso-minimal virtuoso-opensource-6.1-bin virtuoso-opensource-6.1-common xul-ext-ubufox zeitgeist zeitgeist-core zeitgeist-datahub build-essential debhelper dh-apparmor dpkg-dev firefox-globalmenu g++ g++-4.7 html2text kbuild libalgorithm-diff-perl libalgorithm-diff-xs-perl libalgorithm-merge-perl libmail-sendmail-perl libstdc++6-4.7-dev libsys-hostname-long-perl module-assistant openjdk-6-jre openjdk-6-jre-headless openjdk-6-jre-lib po-debconf thunderbird-globalmenu virtualbox-guest-source xchat-indicator" || return 1
+                uninstalldeb "account-plugin-facebook account-plugin-flickr account-plugin-google account-plugin-twitter alacarte appmenu-gtk appmenu-gtk3 appmenu-qt appmenu-qt5 apport apport-symptoms bamfdaemon banshee-extension-soundmenu bison cdparanoia cdrdao compiz compiz-core compiz-gnome compiz-plugins-default curl dconf-tools docbook-xsl flex freepats friends-facebook friends-twitter gir1.2-panelapplet-4.0 gir1.2-rb-3.0 gir1.2-unity-5.0 gnome-applets gnome-applets-data gnome-control-center gnome-control-center-data gnome-control-center-signon gnome-control-center-unity gnome-media gnome-session gnome-session-fallback gnome-user-guide gromit gstreamer0.10-gnomevfs hud humanity-icon-theme icoutils indicator-applet-complete indicator-appmenu indicator-datetime indicator-messages indicator-power indicator-printers indicator-session indicator-sound k3b k3b-data kate-data katepart kde-runtime kde-runtime-data kde-style-oxygen kde-window-manager kde-window-manager-common kdelibs-bin kdelibs5-data kdelibs5-plugins kdoctools kubuntu-debug-installer libattica0.4 libbamf3-1 libbison-dev libcompizconfig0 libdlrestrictions1 libencode-locale-perl libfile-listing-perl libfl-dev libflac++6 libfont-afm-perl libgnome-control-center1 libgnome-media-profiles-3.0-0 libgnome2-canvas-perl libgnome2-perl libgnome2-vfs-perl libgnomevfs2-extra libhtml-form-perl libhtml-format-perl libhtml-parser-perl libhtml-tagset-perl libhtml-tree-perl libhttp-cookies-perl libhttp-daemon-perl libhttp-date-perl libhttp-message-perl libhttp-negotiate-perl libibus-1.0-0 libio-socket-ssl-perl libk3b6 libkactivities-bin libkactivities-models1 libkactivities6 libkatepartinterfaces4 libkcddb4 libkcmutils4 libkde3support4 libkdeclarative5 libkdecorations4abi1 libkdecore5 libkdesu5 libkdeui5 libkdewebkit5 libkdnssd4 libkemoticons4 libkfile4 libkhtml5 libkidletime4 libkio5 libkjsapi4 libkjsembed4 libkmediaplayer4 libknewstuff3-4 libknotifyconfig4 libkntlm4 libkparts4 libkpty4 libkrosscore4 libktexteditor4 libkwineffects1abi4 libkwinglutils1abi1 libkwinnvidiahack4 libkworkspace4abi2 libkxmlrpcclient4 liblwp-mediatypes-perl liblwp-protocol-https-perl libmusicbrainz5-0 libmysqlclient18 libnepomuk4 libnepomukcore4abi1 libnepomukquery4a libnepomukutils4 libnet-http-perl libnet-ssleay-perl libntrack-qt4-1 libntrack0 libnux-4.0-0 libphonon4 libplasma3 libpolkit-qt-1-1 libpoppler-qt4-4 libqapt2 libqapt2-runtime libqca2 libqt4-qt3support libqt4-sql-mysql librhythmbox-core6 libsolid4 libsoprano4 libstreamanalyzer0 libstreams0 libthreadweaver4 libunity-core-6.0-5 libunity-misc4 libunity-webapps0 libvirtodbc0 libwww-perl libwww-robotrules-perl libxcb-damage0 libxml2-utils mysql-common nautilus nepomuk-core nepomuk-core-data notification-daemon ntrack-module-libnl-0 odbcinst odbcinst1debian2 oxygen-icon-theme phonon phonon-backend-gstreamer plasma-scriptengine-javascript python-zeitgeist python3-apport python3-dbus.mainloop.qt python3-distupgrade python3-problem-report python3-pyqt4 python3-sip python3-update-manager qapt-batch rhythmbox rhythmbox-data rhythmbox-mozilla rhythmbox-plugin-cdrecorder rhythmbox-plugin-zeitgeist rhythmbox-plugins rhythmbox-ubuntuone shared-desktop-ontologies soprano-daemon ubiquity-frontend-kde ubuntu-release-upgrader-core unity unity-asset-pool unity-common unity-lens-applications unity-lens-files unity-lens-friends unity-lens-music unity-lens-photos unity-lens-shopping unity-lens-video unity-scope-gdrive unity-scope-musicstores unity-scope-video-remote unity-services unity-webapps-service update-manager-core vcdimager virtuoso-minimal virtuoso-opensource-6.1-bin virtuoso-opensource-6.1-common xul-ext-ubufox zeitgeist zeitgeist-core zeitgeist-datahub build-essential debhelper dh-apparmor dpkg-dev firefox-globalmenu g++ g++-4.7 html2text kbuild libalgorithm-diff-perl libalgorithm-diff-xs-perl libalgorithm-merge-perl libmail-sendmail-perl libstdc++6-4.7-dev libsys-hostname-long-perl module-assistant openjdk-6-jre openjdk-6-jre-headless openjdk-6-jre-lib po-debconf thunderbird-globalmenu virtualbox-guest-source xchat-indicator" || return 1
             fi
 	fi
 
@@ -733,20 +769,26 @@ function _mcos()
 
         #Reset sourcelist
         if [ $BUITSTEP -le 90 ] ; then
-            echo 90 >$BUILDOSSTEP           
-            sudo sh $T/build/core/set_sourcelist.sh $OUTPATH/squashfs-root || return 1
+            echo 90 >$BUILDCOSSTEP           
+            sudo cp $T/build/core/sources.list $OUTPATH/squashfs-root/etc/apt/sources.list.d/official-package-repositories.list
+            sudo cp $T/build/core/cos-keyring_2014.02.21_all.deb $OUTPATH/squashfs-root/tmp/
+            sudo chroot $OUTPATH/squashfs-root /bin/bash -c "dpkg -i /tmp/cdos-keyring_2014.03.07_all.deb" || return 1
+            sudo rm $OUTPATH/squashfs-root/tmp/cos-keyring_2014.02.21_all.deb
+            sudo rm $OUTPATH/squashfs-root/etc/apt/sources.list
+            sudo chroot $OUTPATH/squashfs-root /bin/bash -c "apt-get update" || return 1
         fi        
 
         if [ $BUITSTEP -le 100 ] ; then
             echo 100 >$BUILDOSSTEP
             mountdir  || return 1
-            uninstallmintdeb || return 1
 	    if [ $ISFROMSRC -eq 1 ] ; then
-                uninstalldeb mint-info-xfce || return 1
+                uninstalldebbyapt "libdnet libgadu3 libhal1 libmagickcore5 libmagickwand5 libprelude2 libunwind8 menu imagemagick-common liblqr-1-0" || return 1
             fi
+            uninstallmintdeb || return 1
 	    #wangyu: Debs should be removed by the information of Local Application Group
 		#The cause of umount failure pacakage is "pidgin"
 	    uninstalldeb "cos-meta-codecs libreoffice-base libreoffice-base-core libreoffice-calc libreoffice-emailmerge libreoffice-gnome libreoffice-gtk libreoffice-help-en-gb libreoffice-help-en-us libreoffice-help-zh-cn libreoffice-impress libreoffice-java-common libreoffice-math libreoffice-ogltrans libreoffice-presentation-minimizer libreoffice-writer mythes-en-us banshee gimp gimp-data gimp-help-common gimp-help-en eog transmission-common transmission-gtk brasero vlc vlc-data vlc-nox vlc-plugin-notify vlc-plugin-pulse libvlccore5 libvlc5 brasero-cdrkit brasero-common libbrasero-media3-1" || return 1
+            uninstalldeb "xchat xchat-common" || return 1
 	    if [ $ISFROMSRC -eq 1 ] ; then
                 uninstalldeb "mint-info-xfce banshee-extension-soundmenu" || return 1
             fi
@@ -755,7 +797,7 @@ function _mcos()
             if [ $ISONLINE == 1 ] ; then
                 installdebonline "ubuntu-system-adjustments mint-mdm-themes mint-local-repository mint-flashplugin mint-flashplugin-11 mint-meta-cinnamon mint-meta-core mint-stylish-addon mintdrivers mint-artwork-cinnamon mintsources mintbackup mintstick mintwifi mint-artwork-gnome mint-themes mint-artwork-common mint-backgrounds-olivia mint-x-icons mintsystem mintwelcome mintinstall mintinstall-icons mintnanny mintupdate mintupload mint-info-cinnamon mint-common mint-mirrors mint-translations cinnamon cinnamon-common cinnamon-screensaver nemo nemo-data nemo-share cdos-upgrade"  || return 1
             else
-                installdeb "cinnamon cinnamon-common cinnamon-control-center cinnamon-control-center-data cinnamon-screensaver mint-artwork-cinnamon mint-artwork-common mint-artwork-gnome mint-backgrounds-olivia mintbackup mint-common mintdrivers mint-flashplugin mint-flashplugin-11 mint-info-cinnamon mintinstall mintinstall-icons mint-local-repository mint-mdm-themes mint-meta-core mint-mirrors mintnanny mintsources mintstick mint-stylish-addon mintsystem mint-themes mint-translations mintupdate cdos-upgrade mintupload mintwelcome mintwifi mint-x-icons gir1.2-gtop-2.0 libfcitx-qt5-0 gnome-screenshot gnome-system-monitor libcinnamon-control-center1 nemo nemo-data nemo-share ubuntu-system-adjustments libtimezonemap1 gir1.2-timezonemap-1.0" || return 1
+                installdeb "cinnamon cinnamon-common cinnamon-control-center cinnamon-control-center-data cinnamon-screensaver mint-artwork-cinnamon mint-artwork-common mint-artwork-gnome mint-backgrounds-olivia mintbackup mint-common mintdrivers mint-flashplugin mint-flashplugin-11 mint-info-cinnamon mintinstall mintinstall-icons mint-local-repository mint-mdm-themes mint-meta-core mint-mirrors mintnanny mintsources mintstick mint-stylish-addon mintsystem mint-themes mint-translations mintupdate cdos-upgrade mintupload mintwelcome mintwifi mint-x-icons gir1.2-gtop-2.0 libfcitx-qt5-0 gnome-screenshot gnome-system-monitor libcinnamon-control-center1 nemo nemo-data nemo-share ubuntu-system-adjustments libtimezonemap1 gir1.2-timezonemap-1.0 cdospatchmgr gnome-icon-theme-symbolic" || return 1
             fi
             mountdir || return 1
 	    if [ $ISFROMSRC -eq 1 ] ; then
@@ -832,8 +874,6 @@ VERSION_ID=\"$OSVERSION\"" | sudo tee $OUTPATH/squashfs-root/etc/os-release
             sudo cp $OUT/out/squashfs-root/boot/initrd.img-${KERNEL_VERSION_FULL} $OUT/out/$OSNAME/casper/initrd.lz || return 1
         fi
 
-		
-
         #Install deb by apt-get install 
         if [ $BUITSTEP -le 161 ] ; then
             echo 161 >$BUILDOSSTEP
@@ -863,29 +903,23 @@ VERSION_ID=\"$OSVERSION\"" | sudo tee $OUTPATH/squashfs-root/etc/os-release
             if [ ! -d $OUTPATH/squashfs-root/usr/share/$OSNAME ] ; then
                 mkdir -p $OUTPATH/squashfs-root/usr/share$OSNAME/
             fi
-            echo $NOWTIME>/tmp/buildtime
-            sudo mv /tmp/buildtime $OUTPATH/squashfs-root/usr/share/$OSNAME/buildtime
+            if [ $ISFROMSRC -eq 1 ] ; then
+                echo source $NOWTIME | sudo tee $OUTPATH/squashfs-root/usr/share/cosdesktop/buildtime
+            else
+                echo normal $NOWTIME | sudo tee $OUTPATH/squashfs-root/usr/share/cosdesktop/buildtime
+            fi
             mkiso $ISOFILENAME || return 1
         fi
         echo Finish building $OSFULLNAME.
 
         if [ $BUITSTEP -le 230 ] ; then
-            echo 230 >$BUILDOSSTEP
-            sudo sh $T/build/debug/installkdump.sh $OUTPATH $APPPATH || return 1
+            echo 230 >$BUILDCOSSTEP
+            if [ $IS4DEBUG -eq 1 ] ; then
+                ISODEBUGFILENAME="$ISONAME-debug.iso"
+                mkiso_debug $ISODEBUGFILENAME $OUTPATH $APPPATH || return 1
+                echo Finish building COS Desktop DEBUG.
+            fi
         fi
-
-        if [ $BUITSTEP -le 231 ] ; then
-            echo 231 >$BUILDOSSTEP
-            ISODEBUGFILENAME="$ISONAME-debug.iso"
-            mkiso $ISODEBUGFILENAME || return 1
-        fi
-        echo 231 >$BUILDOSSTEP
-        
-        if [ $BUITSTEP -le 232 ] ; then
-            echo 232 >$BUILDOSSTEP
-            sudo sh $T/build/debug/uninstallkdump.sh $OUTPATH $APPPATH || return 1
-        fi
-        echo Finish building $OSFULLNAME DEBUG.
 
         if [ $BUITSTEP -le 250 ] ; then
             echo 250 >$BUILDOSSTEP
@@ -1056,16 +1090,24 @@ function cclean()
         echo Removing start...
         echo Umounting dir...
         umountdir 2>/dev/null
-	if [ "$?" -ne "0" ] ; then
-	    echo "The device can not be umounted now... Please restart the computer and try it again!"
-	    return 1
-	fi	
+        mount | grep $OUT/out/squashfs-root
+	if [ "$?" -eq "0" ] ; then
+            echo "The device can not be umounted now... Please restart the computer and try it again!"
+            return 1
+            #sudo fuser -k $OUT/out/squashfs-root
+            #umountdir 2>/dev/null
+            #mount | grep $OUT/out/squashfs-root
+            #if [ "$?" -ne "0" ] ; then
+            #    echo "The device can not be umounted now... Please restart the computer and try it again!"
+            #    return 1
+            #fi
+	fi
         for dir in $dirclean
         do
             if [  -e $OUT/$dir ] ; then
-            echo Deleting $OUT/$dir ...
-            sudo rm -r $OUT/$dir
-        fi
+                echo Deleting $OUT/$dir ...
+                sudo rm -rf $OUT/$dir
+            fi
         done
         echo Finished cleaning dir.
     else
@@ -1136,6 +1178,24 @@ function uninstallmintdeb()
         echo "Couldn't locate the top of the tree.  Try setting TOP."
         return 1
     fi
+}
+
+function uninstalldebbyapt()
+{
+    T=$(gettop)
+    if [ ! "$T" ] ; then
+        echo "Couldn't locate the top of the tree.  Try setting TOP."
+        return 1
+    fi
+    if [ $# -lt 1 ] ; then
+        echo Error: no debname param
+        return 1
+    fi
+    if [ ! -e $OUT/out/squashfs-root ] ; then
+        echo Error: No squashfs-root dir exist. Have you executed mcos or uniso once?
+        return 1
+    fi
+    sudo chroot $OUT/out/squashfs-root /bin/bash -c "DEBIAN_FRONTEND=noninteractive apt-get --ignore-missing --yes --force-yes purge $@" || return 1
 }
 
 function uninstalldeb()
